Refactor images, fix blurbs, and add images to VC page
