+++ src/components/About/About.css
@@ -90,4 +90,7 @@
     .for-founders {
         padding-top: 1rem;
     }
+    .incubator-image {
+        width: 60vw;
+    }
 }
