+++ src/images/headshots/vicki.jpeg
